requirements.txt
streamlit>=1.24.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
PyJWT>=2.8.0
google-auth>=2.22.0
